backend/requirements.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.15
decorator==5.2.1
dlib==20.0.0
dnspython==2.7.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Dance==7.1.0
Flask-JWT-Extended==4.7.1
Flask-Session==0.8.0
flatbuffers==25.2.10
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
imutils==0.5.4
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgspec==0.19.0
namex==0.1.0
numpy==2.1.3
oauthlib==3.3.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==311
pyzmq==27.0.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
scipy==1.16.0
six==1.17.0
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.5.0
URLObject==3.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
